+++ .idea/gradle.xml
@@ -13,8 +13,8 @@
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/buildSrc" />
             <option value="$PROJECT_DIR$/di" />
-            <option value="$PROJECT_DIR$/feature-domain-example" />
-            <option value="$PROJECT_DIR$/feature-ui-example" />
+            <option value="$PROJECT_DIR$/feature-domain-intro" />
+            <option value="$PROJECT_DIR$/feature-ui-intro" />
             <option value="$PROJECT_DIR$/navigator" />
             <option value="$PROJECT_DIR$/util-db" />
             <option value="$PROJECT_DIR$/util-kotlin" />
