--- feature-domain-example/build.gradle.kts
@@ -1,44 +1,0 @@
-plugins {
-    id(Dependency.Gradle.LIBRARY)
-    id(Dependency.Gradle.KOTLIN)
-}
-
-android {
-    namespace = "com.notdo.feature_domain_example"
-    compileSdk = Versions.COMPILESDK
-
-    defaultConfig {
-        minSdk = Versions.MINSDK
-        targetSdk = Versions.TARGETSDK
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
-    compileOptions {
-        sourceCompatibility = Versions.JAVAVERSION
-        targetCompatibility = Versions.JAVAVERSION
-    }
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-
-dependencies {
-
-    implementation(Dependency.AndroidX.CORE_KTX)
-    implementation(Dependency.AndroidX.APPCOMPAT)
-    implementation(Dependency.Google.ANDROID_MATERIAL)
-    testImplementation(Dependency.Test.JUNIT)
-    androidTestImplementation(Dependency.Test.ANDROID_JUNIT)
-    androidTestImplementation(Dependency.Test.ESPRESSO)
-}
