+++ feature-domain-intro/src/main/AndroidManifest.xml
