--- feature-ui-example/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="title_activity_main">MainActivity</string>
-</resources>
