+++ settings.gradle
@@ -16,9 +16,9 @@ rootProject.name = "NotDoAos"
 include ':app'
 include ':di'
 include ':navigator'
-include ':feature-ui-example'
-include ':feature-domain-example'
 include ':util-ui'
 include ':util-db'
 include ':util-network'
 include ':util-kotlin'
+include ':feature-ui-intro'
+include ':feature-domain-intro'
